requirements.txt
python-telegram-bot==20.7
openpyxl==3.1.5
python-dotenv==1.0.1
openai==1.35.14
replicate==0.32.1
requests==2.32.3
Pillow==10.4.0
lxml==5.2.2
rapidfuzz==3.9.6
Unidecode==1.3.8 
opencv-python-headless==4.10.0.84
numpy==2.1.1 